move functions to util.go
